Add patches to deal with the new Simple FB / DRM.
Fedora 36 has switched from using the older fbdev and efifb dfivers
in favour of using simplefb instead. This is not immediately
compatible with the NVidia kmod however.

These patches are borrowed from the rpmfusion repo with one edit
to allow the patch to be cleanly applied to the negativo17 build
tree.